rename augur_data model file to data
