+++ collectoss/application/db/models/__init__.py
@@ -1,4 +1,4 @@
-from collectoss.application.db.models.augur_data import (
+from collectoss.application.db.models.data import (
     ChaossMetricStatus,
     ChaossUser,
     ContributorAffiliation,
